+++ openfisca_us/parameters/gov/doe/high_efficiency_electric_home_rebate/elements.yaml
@@ -1,6 +1,6 @@
 description: The US covers this percentage of expenditures for its high-efficiency electric home program rebate program, based on a household's ratio of income to area median income
 values:
-  0000-01-01: {}
+  0000-01-01: []
   2023-01-01:
     - electric_heat_pump_clothes_dryer
     - electric_load_service_center_upgrade
@@ -9,7 +9,7 @@ values:
     - heat_pump_water_heater
     - heat_pump
     - insulation_air_sealing_and_ventilation
-  2032-01-01: {}
+  2032-01-01: []
 metadata:
   unit: program
   name: high_efficiency_electric_home_rebate_elements
