+++ openfisca_us/tests/policy/baseline/gov/doe/high_efficiency_electric_home_rebate_program/capped_electric_stove_cooktop_range_or_oven_rebate.yaml
