+++ openfisca_us/variables/gov/doe/high_efficiency_electric_home_rebate/capped_heat_pump_water_heater_rebate.py
@@ -1,16 +1,16 @@
 from openfisca_us.model_api import *
 
 
-class capped_electric_stove_cooktop_range_or_oven_rebate(Variable):
+class capped_heat_pump_water_heater_rebate(Variable):
     value_type = float
     entity = Household
-    label = "Capped electric stove cooktop range or oven rebate"
+    label = "Capped heat pump water heater rebate"
     documentation = "Before total high efficiency electric home rebate cap"
     definition_period = YEAR
     unit = USD
 
     def formula(household, period, parameters):
         percent_covered = household("high_efficiency_electric_home_rebate_percent_covered", period)
-        expenditures = household("electric_stove_cooktop_range_or_oven_expenditures", period)
-        cap = parameters(period).gov.doe.high_efficiency_electric_home_rebate.cap.electric_stove_cooktop_range_or_oven
+        expenditures = household("heat_pump_water_heater_expenditures", period)
+        cap = parameters(period).gov.doe.high_efficiency_electric_home_rebate.cap.heat_pump_water_heater
         return min_(expenditures * percent_covered, cap)
